+++ packages/apptive_grid_form/lib/src/translation/apptive_grid_translation.dart
@@ -71,4 +71,16 @@ abstract class ApptiveGridTranslation {
 
   /// Message that the form is being submitted
   String get progressSubmitForm;
+
+  /// Hint that the provided email is not valid
+  String get invalidEmail;
+
+  /// Label to clear an input
+  String get clear;
+
+  /// Label to save changes
+  String get save;
+
+  /// Prompt to sign in the field
+  String get signHere;
 }