+++ packages/apptive_grid_form/lib/src/translation/l10n/translation_de.dart
@@ -55,4 +55,12 @@ class ApptiveGridLocalizedTranslation extends ApptiveGridTranslation {
       'Anhänge werden verarbeitet [$processed/$total]';
   @override
   String get progressSubmitForm => "Formular einsenden";
+  @override
+  String get invalidEmail => "Ungültige E-Mail";
+  @override
+  String get clear => "Korrigieren";
+  @override
+  String get save => "Speichern";
+  @override
+  String get signHere => "Hier unterschreiben";
 }