+++ packages/apptive_grid_form/lib/src/translation/l10n/translation_en.dart
@@ -55,4 +55,12 @@ class ApptiveGridLocalizedTranslation extends ApptiveGridTranslation {
       'Processing Attachments [$processed/$total]';
   @override
   String get progressSubmitForm => "Submitting Form";
+  @override
+  String get invalidEmail => "Invalid email";
+  @override
+  String get clear => "Clear";
+  @override
+  String get save => "Save";
+  @override
+  String get signHere => "Sign here";
 }